copy_helper_api/copy-helper-requirements.txt
flask
pillow
redis
python-docx
google-api-python-client
google-auth-oauthlib